Maj du 09.06.20 bis

+++ projet4/index.php
@@ -49,12 +49,14 @@ if (isset($_GET['url']) && !empty($_GET['url']))
     
             
     elseif ($_GET['url']=="blog") {
-        /*var_dump($_POST);*/
+        
         postCheck();
-        deleteCheck();
         listAllPosts();
 
-        
+        if (isset($_POST['delete'])) {
+            deleteCheck();
+              
+        }
     }
     
     elseif ($_GET['url']=="post") {
